feat: add extendible hash
