+++ src/storage/dram/hash_interface.h
@@ -1,12 +1,11 @@
-#ifndef HASH_INTERFACE_H_
-#define HASH_INTERFACE_H_
+#pragma once
 
 #define CAS(_p, _u, _v)                                                        \
   (__atomic_compare_exchange_n(_p, _u, _v, false, __ATOMIC_ACQUIRE,            \
                                __ATOMIC_ACQUIRE))
 
-#include "../util/pair.h"
-#include "../util/timer.h"
+#include "pair.h"
+#include "timer.h"
 
 class Hash {
 public:
@@ -22,5 +21,3 @@ public:
   Timer timer;
   double breakdown = 0;
 };
-
-#endif // _HASH_INTERFACE_H_
