+++ src/storage/kv_engine/kv_engine_register.cc
@@ -3,7 +3,7 @@
 // #include "engine_double_desk.h"
 // #include "engine_dash.h"
 // #include "engine_hashapi.h"
-#include "engine_extendiblehash.h"
+#include "engine_extendible_hash.h"
 #include "engine_map.h"
 // #include "engine_mappm.h"
 // #include "engine_multishard_map_pm.h"
